requirements.txt
fastapi
uvicorn
yfinance
requests
python-dotenv
feedparser
